// frontend/src/api/models/todaysHabitDtoDTO.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API
 * Clean Architecture API with JWT Authentication
 * OpenAPI spec version: v1
 */

export interface TodaysHabitDtoDTO {
  habitId?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  isCompleted?: boolean | null;
  /** @nullable */
  reflection?: string | null;
  /** @nullable */
  completedAt?: string | null;
}